feat: fixed some issues

+++ main_svc/src/main/java/main/server/service/CategoryService.java
@@ -4,6 +4,8 @@ import lombok.extern.slf4j.Slf4j;
 import main.server.dao.Category;
 import main.server.dto.CategoryDto;
 import main.server.dto.NewCategoryDto;
+import main.server.exception.ConflictException;
+import main.server.exception.NotFoundException;
 import main.server.mapper.CategoryMapper;
 import main.server.repository.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,23 +27,46 @@ public class CategoryService {
     public CategoryDto addCategoryAdmin(NewCategoryDto newCategoryDto) {
         log.info("Adding new category: {}", newCategoryDto);
         Category category = CategoryMapper.toCategory(newCategoryDto);
-        categoryRepository.save(category);
-        log.info("Category added: {}", category);
-        return CategoryMapper.toCategoryDto(category);
+        try {
+            Category categoryUp = categoryRepository.save(category);
+            log.info("Category added: {}", categoryUp);
+            return CategoryMapper.toCategoryDto(categoryUp);
+        } catch (Exception e) {
+            log.error("Error adding category: {}", e.getMessage());
+            throw new ConflictException(e.getMessage());
+        }
     }
 
     public void deleteCategoryAdmin(int catId) {
         log.info("Deleting category with id: {}", catId);
-        categoryRepository.deleteById(catId);
+        if (!categoryRepository.existsById(catId)) {
+            log.error("Category not found with id: {}", catId);
+            throw new NotFoundException("Category not found");
+        }
+        try {
+            categoryRepository.deleteById(catId);
+        } catch (Exception e) {
+            log.error("Error deleting category: {}", e.getMessage());
+            throw new ConflictException(e.getMessage());
+        }
         log.info("Category deleted with id: {}", catId);
     }
     public CategoryDto updateCategoryAdmin(int catId, CategoryDto newCategoryDto) {
         log.info("Updating category: {}", newCategoryDto);
+        if (!categoryRepository.existsById(catId)) {
+            log.error("Category not found with id: {}", catId);
+            throw new NotFoundException("Category not found");
+        }
         Category category = CategoryMapper.toCategory(newCategoryDto);
         category.setId(catId);
-        categoryRepository.save(category);
-        log.info("Category updated: {}", category);
-        return CategoryMapper.toCategoryDto(category);
+        try{
+            Category categoryUp = categoryRepository.save(category);
+            log.info("Category updated: {}", categoryUp);
+            return CategoryMapper.toCategoryDto(categoryUp);
+        } catch (Exception e) {
+            log.error("Error updating category: {}", e.getMessage());
+            throw new ConflictException(e.getMessage());
+        }
     }
 
     public List<CategoryDto> getCategoriesPub() {
@@ -52,7 +77,7 @@ public class CategoryService {
     }
     public CategoryDto getCategoryByIdPub(int catId) {
         log.info("Getting category with id: {}", catId);
-        Category category = categoryRepository.findById(catId).orElseThrow();
+        Category category = categoryRepository.findById(catId).orElseThrow( () -> new NotFoundException("Category with id=" + catId +" was not found"));
         log.info("Found category: {}", category);
         return CategoryMapper.toCategoryDto(category);
     }
